+++ tests/integration_tests/expression.rs
@@ -8,8 +8,8 @@ fn test_scalarspec() {
   <text xy="^" text="{{#a~w}}"/>
 "#;
     let expected = r#"
-  <rect id="a" width="20" height="20"/>
-  <text x="10" y="10" class="d-text">20</text>
+<rect id="a" width="20" height="20"/>
+<text x="10" y="10" class="d-text">20</text>
 "#;
 
     assert_eq!(
@@ -23,8 +23,8 @@ fn test_scalarspec() {
   <text xy="^" text="{{^~w}}"/>
 "#;
     let expected = r#"
-  <rect width="20" height="20"/>
-  <text x="10" y="10" class="d-text">20</text>
+<rect width="20" height="20"/>
+<text x="10" y="10" class="d-text">20</text>
 "#;
 
     assert_eq!(
@@ -38,8 +38,8 @@ fn test_scalarspec() {
   <text xy="^" text="{{#a-1~w}}"/>
 "#;
     let expected = r#"
-  <rect id="a-1" width="20" height="20"/>
-  <text x="10" y="10" class="d-text">20</text>
+<rect id="a-1" width="20" height="20"/>
+<text x="10" y="10" class="d-text">20</text>
 "#;
 
     assert_eq!(