+++ tests/integration_tests/indent.rs
@@ -3,7 +3,7 @@ use svgdx::transform_str_default;
 #[test]
 fn test_indent_ws_only() {
     let input = r#"    "#;
-    let expected = r#"    "#;
+    let expected = r#""#;
     let output = transform_str_default(input).unwrap();
     assert_eq!(output, expected);
 }
@@ -14,8 +14,7 @@ fn test_indent_none() {
 <rect xy="10 20" wh="20 60" />
 <rect id="z" xy="25 35" wh="^" />
 "#;
-    let expected = r#"
-<rect x="10" y="20" width="20" height="60"/>
+    let expected = r#"<rect x="10" y="20" width="20" height="60"/>
 <rect id="z" x="25" y="35" width="20" height="60"/>
 "#;
     let output = transform_str_default(input).unwrap();
@@ -24,14 +23,12 @@ fn test_indent_none() {
 
 #[test]
 fn test_indent_constant() {
-    // OOPS: looks like any leading indent / text event is dropped...
     let input = r#"
   <rect xy="10 20" wh="20 60" />
   <rect id="z" xy="25 35" wh="^" />
 "#;
-    let expected = r#"
-  <rect x="10" y="20" width="20" height="60"/>
-  <rect id="z" x="25" y="35" width="20" height="60"/>
+    let expected = r#"<rect x="10" y="20" width="20" height="60"/>
+<rect id="z" x="25" y="35" width="20" height="60"/>
 "#;
     let output = transform_str_default(input).unwrap();
     assert_eq!(output, expected);
@@ -40,9 +37,8 @@ fn test_indent_constant() {
         <rect xy="10 20" wh="20 60" />
         <rect id="z" xy="25 35" wh="^" />
 "#;
-    let expected = r#"
-        <rect x="10" y="20" width="20" height="60"/>
-        <rect id="z" x="25" y="35" width="20" height="60"/>
+    let expected = r#"<rect x="10" y="20" width="20" height="60"/>
+<rect id="z" x="25" y="35" width="20" height="60"/>
 "#;
     let output = transform_str_default(input).unwrap();
     assert_eq!(output, expected);
@@ -55,10 +51,9 @@ fn test_indent_out_of_order() {
   <rect id="a" xy="1" wh="1"/>
 "##;
 
-    let expected = r##"
-  <line x1="1" y1="1.5" x2="2" y2="1.5"/>
-  <text x="1.5" y="1.5" class="d-text">a</text>
-  <rect id="a" x="1" y="1" width="1" height="1"/>
+    let expected = r##"<line x1="1" y1="1.5" x2="2" y2="1.5"/>
+<text x="1.5" y="1.5" class="d-text">a</text>
+<rect id="a" x="1" y="1" width="1" height="1"/>
 "##;
 
     let output = transform_str_default(input).unwrap();
@@ -72,11 +67,20 @@ fn test_indent_ooo_varying() {
    <rect id="a" xy="1" wh="1"/>
 "##;
 
-    let expected = r##"
-     <line x1="1" y1="1.5" x2="2" y2="1.5"/>
-     <text x="1.5" y="1.5" class="d-text">a</text>
-   <rect id="a" x="1" y="1" width="1" height="1"/>
+    let expected = r##"<line x1="1" y1="1.5" x2="2" y2="1.5"/>
+<text x="1.5" y="1.5" class="d-text">a</text>
+<rect id="a" x="1" y="1" width="1" height="1"/>
 "##;
+
+    let output = transform_str_default(input).unwrap();
+    assert_eq!(output, expected);
+}
+
+#[test]
+fn test_indent_same_line_run() {
+    let input = r#"<rect xy="0" wh="1"/><rect xy="^|h" wh="1"/>"#;
+    let expected =
+        r#"<rect x="0" y="0" width="1" height="1"/><rect x="1" y="0" width="1" height="1"/>"#;
 
     let output = transform_str_default(input).unwrap();
     assert_eq!(output, expected);